changed the vim handling
This change represents a complete cleanup of my vim setup of the last 4
years. It's not perfect yet, but it's much more cleaner on what the
config is intended to do.

I pulled in some git projects for syntax highlighting on haml, scss and
ruby. Maybe I will pull them out into git submodules, but at the moment,
it's okay that way.

I used the hints from http://mislav.uniqpath.com/2011/12/vim-revisited/
to rebuild the config.